Use int64 for ranges. Doesn't matter now, but it will with chunking.

+++ byterange.go
@@ -10,9 +10,9 @@ func RangeKey(key string, start, end int64) string {
 	return fmt.Sprintf("%s$$%d-%d", key, start, end)
 }
 
-func KeyToRange(rk string) (string, int, int, bool) {
+func KeyToRange(rk string) (string, int64, int64, bool) {
 	var key string
-	var start, end int
+	var start, end int64
 	if n, err := fmt.Sscanf(key, "%s$$%d-%d", &key, &start, &end); err != nil && n == 3 {
 		key = strings.ReplaceAll(key, "\\$", "$")
 		return key, start, end, true
